Deconstructed
Turned some of its JS functions and CSS styles into separate components to also be used in the login page.
