+++ js/modules/errors/closeErrModal.js
@@ -1,0 +1,6 @@
+function closeErrModal() {
+  const errContainer = document.querySelector(".err-container");
+  errContainer.classList.add("closed");
+}
+
+export default closeErrModal;
