+++ js/modules/errors/displayErrTag.js
@@ -1,0 +1,14 @@
+/**
+ * Creates an error p tag and makes it the sibling of the provided input
+ * @param {*} inputTag
+ * @param {*} message
+ */
+function displayErrTag(inputTag, message) {
+  const errTag = document.createElement("p");
+  errTag.classList.add("err");
+  errTag.innerText = message;
+
+  inputTag.insertAdjacentElement("afterend", errTag);
+}
+
+export default displayErrTag;
