+++ signUp.html
@@ -4,6 +4,7 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>Document</title>
     <link rel="stylesheet" href="/styles/signUp_styles.css" />
+    <link rel="stylesheet" href="/styles/err_style.css" />
     <script src="/js/signUp.js" type="module" defer></script>
   </head>
   <body>
