+++ styles/err_style.css
@@ -1,0 +1,90 @@
+/*Styless for the error modal and the input errors on the signUp and logIn pages*/
+
+/*Select and Input tags*/
+input.err {
+  border: 2px solid #b70202;
+}
+
+p.err {
+  color: #b70202;
+  margin: -6px 0px 0px 0px;
+  text-align: justify;
+}
+
+p.err::before {
+  content: "* ";
+}
+
+select.err {
+  border: 2px solid #b70202;
+}
+
+/*Error Modal---*/
+.err-container {
+  position: fixed;
+  bottom: 5px;
+  width: 100%;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+}
+
+.err-container.closed {
+  display: none;
+}
+
+.err-modal {
+  border: 2px solid #da0c0c;
+  border-radius: 12px;
+  background-color: #c70303;
+  padding: 5px;
+  height: max-content;
+  width: 50%;
+
+  position: relative;
+}
+
+.err-modal p {
+  color: white;
+  font-weight: 700;
+  text-align: center;
+}
+
+.err-modal button {
+  position: absolute;
+  right: 8px;
+  top: 0px;
+
+  border: none;
+  background: none;
+  color: #fff;
+  font-weight: 900;
+  font-size: 1.3rem;
+  cursor: pointer;
+  margin: 0px;
+  padding: 0px;
+}
+
+/*Responsive-----*/
+@media (width <=700px) {
+  .err-container {
+    top: 5px;
+    bottom: revert;
+  }
+
+  .err-modal {
+    padding-top: 10px;
+    height: max-content;
+    width: 95%;
+  }
+
+  .err-modal button {
+    right: 8px;
+    top: 0px;
+
+    font-size: 1.3rem;
+
+    margin: 0px;
+    padding: 0px;
+  }
+}
